+++ .gitattributes
@@ -1,3 +1,4 @@
+* text=auto eol=lf
 /gradlew text eol=lf
 *.bat text eol=crlf
 *.jar binary
